+++ src/Test/TestCase.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Hirtz\Media\Test;
 
-use Hirtz\Media\Models\Collections\FolderCollection;
 use Override;
 
 class TestCase extends \Hirtz\Skeleton\Test\TestCase
@@ -14,14 +13,5 @@ class TestCase extends \Hirtz\Skeleton\Test\TestCase
     {
         $this->config ??= require(__DIR__ . '/../../config/test.php');
         parent::setUp();
-
-        FolderCollection::reset();
-    }
-
-    #[\Override]
-    protected function tearDown(): void
-    {
-        FolderCollection::reset();
-        parent::tearDown();
     }
 }